cambiar mil cosas la neta no me acuerdo todas

+++ database/migrations/2019_11_25_144840_alter_reservations_table.php
@@ -13,12 +13,6 @@ class AlterReservationsTable extends Migration
      */
     public function up()
     {
-        Schema::table('reservations', function (Blueprint $table) {
-            $table->dropColumn('start_date');
-            $table->dropColumn('end_date');
-            $table->datetime('start_datetime')->nullable();
-            $table->datetime('end_datetime')->nullable();
-        });
     }
 
     /**
@@ -28,11 +22,11 @@ class AlterReservationsTable extends Migration
      */
     public function down()
     {
-        Schema::table('reservations', function (Blueprint $table) {
-            $table->dropColumn('start_datetime');
-            $table->dropColumn('end_datetime');
-            $table->date('start_date')->nullable();
-            $table->date('end_date')->nullable();
-        });
+        if (Schema::hasColumn('reservations', 'start_datetime')) {
+            Schema::table('reservations', function (Blueprint $table) {
+                $table->dropColumn('start_datetime');
+                $table->dropColumn('end_datetime');
+            });
+        }
     }
 }
